# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Observatorio ULEAM API",
        "description": "API para el Sistema de Gestión y Visualización de Datos Universitarios",
        "contact": {
            "name": "ULEAM",
            "email": "soporte@uleam.edu.ec"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/datasets": {
            "get": {
                "tags": [
                    "Datasets"
                ],
                "summary": "Listar datasets",
                "description": "Obtiene la lista de datasets del usuario",
                "operationId": "ad0a3cd4988a8e12209a2a4756cb92e1",
                "parameters": [
                    {
                        "name": "departamento_id",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de datasets"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Datasets"
                ],
                "summary": "Crear dataset",
                "description": "Crea un nuevo dataset y sube el archivo Excel",
                "operationId": "911ede3cc894bd84f2b6aac27436947d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "departamento_id",
                                    "nombre",
                                    "archivo"
                                ],
                                "properties": {
                                    "departamento_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "archivo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dataset creado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/datasets/{id}": {
            "get": {
                "tags": [
                    "Datasets"
                ],
                "summary": "Obtener dataset",
                "description": "Obtiene un dataset específico con sus metadatos",
                "operationId": "a04d3c641c3509d069d2ff5feb87a615",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dataset encontrado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Datasets"
                ],
                "summary": "Eliminar dataset",
                "description": "Elimina un dataset y todos sus datos",
                "operationId": "56215870ed6a4a21cd6908bceb6d7ef0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dataset eliminado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/datasets/{id}/import": {
            "post": {
                "tags": [
                    "Datasets"
                ],
                "summary": "Analizar archivo Excel",
                "description": "Analiza el archivo Excel y detecta los tipos de columnas",
                "operationId": "c8253d0ac975356a8c7b6211508b99ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Análisis completado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "columnas": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "preview": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/datasets/{id}/confirm": {
            "post": {
                "tags": [
                    "Datasets"
                ],
                "summary": "Confirmar importación",
                "description": "Confirma los tipos de datos y procesa la importación completa",
                "operationId": "26181625f19ba3733ab90b0aab3347a2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "columnas"
                                ],
                                "properties": {
                                    "columnas": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "nombre_columna": {
                                                    "type": "string"
                                                },
                                                "tipo_dato": {
                                                    "type": "string",
                                                    "enum": [
                                                        "NUMERICO",
                                                        "CATEGORICO",
                                                        "FECHA",
                                                        "TEXTO"
                                                    ]
                                                },
                                                "es_visible": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Importación completada"
                    },
                    "400": {
                        "description": "Error en la importación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Ingresa el token en formato: Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}